+++ src/commands/auth-export.js
@@ -4,11 +4,11 @@ var clc = require("cli-color");
 var fs = require("fs");
 var os = require("os");
 
-var Command = require("../command");
+var { Command } = require("../command");
 var accountExporter = require("../accountExporter");
 var getProjectId = require("../getProjectId");
 var logger = require("../logger");
-var requirePermissions = require("../requirePermissions");
+var { requirePermissions } = require("../requirePermissions");
 
 var MAX_BATCH_SIZE = 1000;