+++ src/commands/auth-import.js
@@ -6,11 +6,11 @@ var fs = require("fs");
 var jsonStream = require("JSONStream");
 var _ = require("lodash");
 
-var Command = require("../command");
+var { Command } = require("../command");
 var accountImporter = require("../accountImporter");
 var getProjectId = require("../getProjectId");
 var logger = require("../logger");
-var requirePermissions = require("../requirePermissions");
+var { requirePermissions } = require("../requirePermissions");
 var utils = require("../utils");
 
 var MAX_BATCH_SIZE = 1000;