+++ src/commands/ext-info.ts
@@ -1,12 +1,12 @@
 import * as clc from "cli-color";
 import * as _ from "lodash";
 
-import * as Command from "../command";
+import { Command } from "../command";
 import { resolveSource } from "../extensions/resolveSource";
 import * as extensionsApi from "../extensions/extensionsApi";
 import { ensureExtensionsApiEnabled, logPrefix } from "../extensions/extensionsHelper";
 import * as logger from "../logger";
-import * as requirePermissions from "../requirePermissions";
+import { requirePermissions } from "../requirePermissions";
 import * as utils from "../utils";
 
 import * as marked from "marked";