+++ src/commands/ext-list.ts
@@ -1,8 +1,8 @@
-import * as Command from "../command";
+import { Command } from "../command";
 import * as getProjectId from "../getProjectId";
 import { listExtensions } from "../extensions/listExtensions";
 import { ensureExtensionsApiEnabled } from "../extensions/extensionsHelper";
-import * as requirePermissions from "../requirePermissions";
+import { requirePermissions } from "../requirePermissions";
 
 module.exports = new Command("ext:list")
   .description("list all the extensions that are installed in your Firebase project")