+++ src/commands/functions-config-get.js
@@ -1,10 +1,10 @@
 "use strict";
 
 var _ = require("lodash");
-var Command = require("../command");
+var { Command } = require("../command");
 var getProjectId = require("../getProjectId");
 var logger = require("../logger");
-var requirePermissions = require("../requirePermissions");
+var { requirePermissions } = require("../requirePermissions");
 var functionsConfig = require("../functionsConfig");
 
 function _materialize(projectId, path) {