+++ src/commands/functions-config-set.js
@@ -2,9 +2,9 @@
 
 var clc = require("cli-color");
 
-var Command = require("../command");
+var { Command } = require("../command");
 var getProjectId = require("../getProjectId");
-var requirePermissions = require("../requirePermissions");
+var { requirePermissions } = require("../requirePermissions");
 var logger = require("../logger");
 var utils = require("../utils");
 var functionsConfig = require("../functionsConfig");