+++ src/commands/functions-shell.js
@@ -1,9 +1,9 @@
 "use strict";
 
-var Command = require("../command");
-var requirePermissions = require("../requirePermissions");
-var requireConfig = require("../requireConfig");
+var { Command } = require("../command");
+var { requirePermissions } = require("../requirePermissions");
 var action = require("../functionsShellCommandAction");
+var requireConfig = require("../requireConfig");
 
 module.exports = new Command("functions:shell")
   .description("launch full Node shell with emulated functions")