+++ src/commands/setup-emulators-database.js
@@ -1,10 +1,12 @@
 "use strict";
 
-const Command = require("../command");
+const { Command } = require("../command");
 const downloadEmulator = require("../emulator/download");
 
-const name = "database";
+const NAME = "database";
 
-module.exports = new Command("setup:emulators:" + name)
-  .description("downloads the " + name + " emulator")
-  .action(downloadEmulator.bind(this, name));
+module.exports = new Command(`setup:emulators:${NAME}`)
+  .description(`downloads the ${NAME} emulator`)
+  .action((options) => {
+    return downloadEmulator(NAME);
+  });