+++ src/test/emulators/functionsEmulator.spec.ts
@@ -1,12 +1,9 @@
 import { expect } from "chai";
-import { FunctionsEmulator, FunctionsRuntimeInstance } from "../../emulator/functionsEmulator";
+import { FunctionsEmulator, InvokeRuntimeOpts } from "../../emulator/functionsEmulator";
 import * as supertest from "supertest";
-import { FunctionRuntimeBundles, TIMEOUT_LONG, TIMEOUT_MED } from "./fixtures";
+import { TIMEOUT_LONG, MODULE_ROOT } from "./fixtures";
 import * as logger from "../../logger";
-import {
-  EmulatedTriggerType,
-  FunctionsRuntimeBundle,
-} from "../../emulator/functionsEmulatorShared";
+import { EmulatedTriggerType } from "../../emulator/functionsEmulatorShared";
 import * as express from "express";
 import { RuntimeWorker } from "../../emulator/functionsRuntimeWorker";
 
@@ -20,21 +17,26 @@ if ((process.env.DEBUG || "").toLowerCase().indexOf("spec") >= 0) {
 }
 
 const functionsEmulator = new FunctionsEmulator({
-  projectId: "",
-  functionsDir: "",
+  projectId: "fake-project-id",
+  functionsDir: MODULE_ROOT,
 });
-const startFunctionRuntime = functionsEmulator.startFunctionRuntime;
+
+// This is normally discovered in FunctionsEmulator#start()
+functionsEmulator.nodeBinary = process.execPath;
+
+// TODO(samstern): This is an ugly way to just override the InvokeRuntimeOpts on each call
+const startFunctionRuntime = functionsEmulator.startFunctionRuntime.bind(functionsEmulator);
 function UseFunctions(triggers: () => {}): void {
   const serializedTriggers = triggers.toString();
 
   functionsEmulator.startFunctionRuntime = (
-    bundleTemplate: FunctionsRuntimeBundle,
     triggerId: string,
     triggerType: EmulatedTriggerType,
-    nodeBinary: string,
-    proto?: any
+    proto?: any,
+    runtimeOpts?: InvokeRuntimeOpts
   ): RuntimeWorker => {
-    return startFunctionRuntime(bundleTemplate, triggerId, triggerType, nodeBinary, proto, {
+    return startFunctionRuntime(triggerId, triggerType, proto, {
+      nodeBinary: process.execPath,
       serializedTriggers,
     });
   };
@@ -53,9 +55,7 @@ describe("FunctionsEmulator-Hub", () => {
       };
     });
 
-    await supertest(
-      functionsEmulator.createHubServer(FunctionRuntimeBundles.template, process.execPath)
-    )
+    await supertest(functionsEmulator.createHubServer())
       .get("/fake-project-id/us-central1/function_id")
       .expect(200)
       .then((res) => {
@@ -75,9 +75,7 @@ describe("FunctionsEmulator-Hub", () => {
       };
     });
 
-    await supertest(
-      functionsEmulator.createHubServer(FunctionRuntimeBundles.template, process.execPath)
-    )
+    await supertest(functionsEmulator.createHubServer())
       .get("/fake-project-id/us-central1/function_id/")
       .expect(200)
       .then((res) => {
@@ -97,14 +95,28 @@ describe("FunctionsEmulator-Hub", () => {
       };
     });
 
-    await supertest(
-      functionsEmulator.createHubServer(FunctionRuntimeBundles.template, process.execPath)
-    )
+    await supertest(functionsEmulator.createHubServer())
       .get("/fake-project-id/us-central1/function_id/a/b")
       .expect(200)
       .then((res) => {
         expect(res.body.path).to.deep.equal("/a/b");
       });
+  }).timeout(TIMEOUT_LONG);
+
+  it("should reject requests to a non-emulator path", async () => {
+    UseFunctions(() => {
+      return {
+        function_id: require("firebase-functions").https.onRequest(
+          (req: express.Request, res: express.Response) => {
+            res.json({ path: req.path });
+          }
+        ),
+      };
+    });
+
+    await supertest(functionsEmulator.createHubServer())
+      .get("/foo/bar/baz")
+      .expect(404);
   }).timeout(TIMEOUT_LONG);
 
   it("should rewrite req.path to hide /:project_id/:region/:trigger_id", async () => {
@@ -119,9 +131,7 @@ describe("FunctionsEmulator-Hub", () => {
       };
     });
 
-    await supertest(
-      functionsEmulator.createHubServer(FunctionRuntimeBundles.template, process.execPath)
-    )
+    await supertest(functionsEmulator.createHubServer())
       .get("/fake-project-id/us-central1/function_id/sub/route/a")
       .expect(200)
       .then((res) => {
@@ -141,9 +151,7 @@ describe("FunctionsEmulator-Hub", () => {
       };
     });
 
-    await supertest(
-      functionsEmulator.createHubServer(FunctionRuntimeBundles.template, process.execPath)
-    )
+    await supertest(functionsEmulator.createHubServer())
       .get("/fake-project-id/us-central1/function_id/sub/route/a")
       .expect(200)
       .then((res) => {
@@ -163,9 +171,7 @@ describe("FunctionsEmulator-Hub", () => {
       };
     });
 
-    await supertest(
-      functionsEmulator.createHubServer(FunctionRuntimeBundles.template, process.execPath)
-    )
+    await supertest(functionsEmulator.createHubServer())
       .post("/fake-project-id/us-central1/function_id/sub/route/a")
       .send({ hello: "world" })
       .expect(200)
@@ -186,9 +192,7 @@ describe("FunctionsEmulator-Hub", () => {
       };
     });
 
-    await supertest(
-      functionsEmulator.createHubServer(FunctionRuntimeBundles.template, process.execPath)
-    )
+    await supertest(functionsEmulator.createHubServer())
       .get("/fake-project-id/us-central1/function_id/sub/route/a?hello=world")
       .expect(200)
       .then((res) => {