+++ src/test/emulators/functionsEmulatorRuntime.spec.ts
@@ -1,14 +1,20 @@
 import { expect } from "chai";
-import { invokeRuntime, InvokeRuntimeOpts } from "../../emulator/functionsEmulator";
+import { InvokeRuntimeOpts, FunctionsEmulator } from "../../emulator/functionsEmulator";
 import { EmulatorLog } from "../../emulator/types";
 import { FunctionsRuntimeBundle } from "../../emulator/functionsEmulatorShared";
 import { RuntimeWorker } from "../../emulator/functionsRuntimeWorker";
 import { Change } from "firebase-functions";
 import { DocumentSnapshot } from "firebase-functions/lib/providers/firestore";
-import { FunctionRuntimeBundles, TIMEOUT_LONG, TIMEOUT_MED } from "./fixtures";
+import { FunctionRuntimeBundles, TIMEOUT_LONG, TIMEOUT_MED, MODULE_ROOT } from "./fixtures";
 import * as request from "request";
 import * as express from "express";
 import * as _ from "lodash";
+
+const functionsEmulator = new FunctionsEmulator({
+  projectId: "fake-project-id",
+  functionsDir: MODULE_ROOT,
+});
+functionsEmulator.nodeBinary = process.execPath;
 
 async function _countLogEntries(worker: RuntimeWorker): Promise<{ [key: string]: number }> {
   const runtime = worker.runtime;
@@ -29,11 +35,11 @@ function InvokeRuntimeWithFunctions(
 ): RuntimeWorker {
   const serializedTriggers = triggers.toString();
 
-  opts = opts || {};
+  opts = opts || { nodeBinary: process.execPath };
   opts.ignore_warnings = true;
   opts.serializedTriggers = serializedTriggers;
 
-  return invokeRuntime(process.execPath, frb, opts);
+  return functionsEmulator.invokeRuntime(frb, opts);
 }
 
 /**
@@ -460,6 +466,7 @@ describe("FunctionsEmulator-Runtime", () => {
             };
           },
           {
+            nodeBinary: process.execPath,
             env: {
               CLOUD_RUNTIME_CONFIG: JSON.stringify({
                 does: { exist: "already exists" },