+++ CHANGELOG.md
@@ -1,1 +1,4 @@
-* Updated Extensions registry address.
+* Fixes a bug where the Functions emulator ignored the "host" configuration (#1722)
+* Fixes a bug where the Functions emulator accepted requests to too many paths (#1773)
+* Adds printing of Firebase console links for Firebase Extensions after operations.
+* Updated Firebase Extensions registry address.